+++ src/include/ops/tf.h
@@ -72,13 +72,13 @@ struct vaccel_tf_node {
 	char *name;
 
         /* id of the node */
-        int64_t id;
+        long long int id;
 };
 
-struct vaccel_tf_node *vaccel_tf_node_new(const char *name, int64_t id);
+struct vaccel_tf_node *vaccel_tf_node_new(const char *name, long long int id);
 void vaccel_tf_node_destroy(struct vaccel_tf_node *node);
 const char *vaccel_tf_node_get_name(struct vaccel_tf_node *node);
-int64_t vaccel_tf_node_get_id(struct vaccel_tf_node *node);
+long long int vaccel_tf_node_get_id(struct vaccel_tf_node *node);
 
 
 struct vaccel_tf_tensor {
@@ -93,7 +93,7 @@ struct vaccel_tf_tensor {
 
         /* dimensions of the data */
         int nr_dims;
-        int64_t *dims;
+        long long int *dims;
 
         /* Data type */
         enum vaccel_tf_data_type data_type;
@@ -102,13 +102,13 @@ struct vaccel_tf_tensor {
 struct vaccel_tf_tensor *
 vaccel_tf_tensor_new(
 	int nr_dims,
-	int64_t *dims,
+	long long int *dims,
 	enum vaccel_tf_data_type type
 );
 
 struct vaccel_tf_tensor *
 vaccel_tf_tensor_allocate(
-	int nr_dims, int64_t *dims,
+	int nr_dims, long long int *dims,
 	enum vaccel_tf_data_type type,
 	size_t total_size
 );
